backend/requirements.txt
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
curl_cffi==0.13.0
Django==5.2.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
frozendict==2.4.6
idna==3.10
multitasking==0.0.12
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.2
peewee==3.18.2
platformdirs==4.4.0
joblib==1.4.2
scikit-learn==1.5.2
protobuf==6.32.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yfinance==0.2.65
